this types of shortcut only makes sense when tabs are visible, so better to co-locate with ui component
