+++ apps/desktop/src/routes/app/main/_layout.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute, Outlet, useRouteContext } from "@tanstack/react-router";
-import { useCallback, useEffect } from "react";
+import { useEffect } from "react";
 
 import { toolFactories } from "../../../chat/tools";
 import { useSearchEngine } from "../../../contexts/search/engine";
@@ -17,7 +17,7 @@ export const Route = createFileRoute("/app/main/_layout")({
 
 function Component() {
   const { persistedStore, internalStore } = useRouteContext({ from: "__root__" });
-  const { registerOnClose, registerOnEmpty, currentTab, openNew, close } = useTabs();
+  const { registerOnClose, registerOnEmpty, currentTab, openNew } = useTabs();
 
   useEffect(() => {
     return registerOnClose((tab) => {
@@ -49,44 +49,17 @@ function Component() {
     return registerOnEmpty(createDefaultSession);
   }, [currentTab, persistedStore, internalStore, registerOnEmpty, openNew]);
 
-  const handleNewTab = useCallback((closeCurrentFirst: boolean) => {
-    const user_id = internalStore?.getValue("user_id");
-    const sessionId = id();
-    persistedStore?.setRow("sessions", sessionId, {
-      user_id,
-      created_at: new Date().toISOString(),
-      title: "",
-    });
-
-    if (closeCurrentFirst && currentTab) {
-      close(currentTab);
-    }
-
-    openNew({
-      type: "sessions",
-      id: sessionId,
-      active: true,
-      state: { editor: "raw" },
-    });
-  }, [persistedStore, internalStore, currentTab, close, openNew]);
-
   return (
     <SearchEngineProvider store={persistedStore}>
       <SearchUIProvider>
-        <SearchShortcutBridge onNewTab={handleNewTab} />
+        <ShellProvider>
+          <ToolRegistryProvider>
+            <ToolRegistration />
+            <Outlet />
+          </ToolRegistryProvider>
+        </ShellProvider>
       </SearchUIProvider>
     </SearchEngineProvider>
-  );
-}
-
-function SearchShortcutBridge({ onNewTab }: { onNewTab: (closeCurrentFirst: boolean) => void }) {
-  return (
-    <ShellProvider onNewTab={onNewTab}>
-      <ToolRegistryProvider>
-        <ToolRegistration />
-        <Outlet />
-      </ToolRegistryProvider>
-    </ShellProvider>
   );
 }
